refact: currenttimemillis -> nanotime

+++ src/main/kotlin/taskforce/fcfs/clientqueue/result/JoinResult.kt
@@ -2,6 +2,6 @@ package taskforce.fcfs.clientqueue.result
 
 
 sealed class JoinResult {
-    data class Success(val rank: Int, val joinTime: Double): JoinResult()
+    data class Success(val rank: Int, val joinTime: Long): JoinResult()
     data class Fail(val message: String): JoinResult()
 }
